simplify iteration on P

+++ cpp/include/games/stochastic_nim/Constants.hpp
@@ -8,7 +8,7 @@ namespace stochastic_nim {
 
 constexpr int kNumPlayers = 2;
 constexpr int kMaxStonesToTake = 3;
-constexpr unsigned int kStartingStones = 21;
+constexpr int kStartingStones = 21;
 
 // After each player's turn, a random int r is chose with probability kChanceEventProbs[r],
 // and r stones are removed from the pile.
@@ -25,5 +25,5 @@ constexpr core::action_mode_t kChanceMode = 1;
 constexpr core::seat_index_t kPlayer0 = 0;
 constexpr core::seat_index_t kPlayer1 = 1;
 
-constexpr int kStartingStonesBitWidth = std::bit_width(kStartingStones);
+constexpr int kStartingStonesBitWidth = std::bit_width(static_cast<unsigned int>(kStartingStones));
 }
